requirements.txt
Flask
spacy
transformers
scikit-learn
nltk
PyPDF2
python-docx
flask-wtf
torch  
fpdf